HACK: OP6: cfg80211: Remove all WARN_ONs in cfg80211_calculate_bitrate*
This is super bad but OnePlus/CAF's Wi-Fi driver is clearly triggering
these and compiling then replacing it is really intrusive (requires
disabling AVB on /vendor and pushing the module, requiring a full ROM
flash to uninstall the kernel).

Along the same lines as commit c00da6cb6bd2 ("Revert "cfg80211: add
warning when calculating MCS rates >= 32"").

This is probably not necessary on your device, I'm going to laugh if you
blindly pick this :)

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Anirudh Gupta <anirudhgupta109@gmail.com>